requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
alluka==0.3.3
attrs==24.3.0
Brotli==1.1.0
cffi==1.17.1
ciso8601==2.3.2
colorlog==6.9.0
frozenlist==1.5.0
hikari==2.1.0
hikari-arc==1.4.0
idna==3.10
multidict==6.1.0
orjson==3.10.12
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
python-dotenv==1.0.1
ruff==0.8.3
typing_extensions==4.12.2
uvloop==0.21.0
yarl==1.18.3
